gblink_pinconf: create new unit for handling pinconf
Split off from gblink functions. Will need to create a handful of
different and more advanced pinconf handling functions so it is
going to be ideal in the long run to split this out.

Moves the struct gblink definition to gblink_i.h so the pinconf
unit can directly access parts of the struct.

Signed-off-by: Kris Bahnsen <Kris@KBEmbedded.com>