Some attempts at implementation, deferred till later
After thinking about this more, printer send support needs to go on
hold for a bit. I need to re-architect some things so that the send
can happen asynchronously. Since a complete send needs multiple packets,
it would currently need to busyloop for the entire thing.

I could architect the send process such that the application wanting
to send data would need to sned a bunch of packets and handle the
everything itself. That might be the better idea long-term.

However, first, there are some more wrinkles to work out in gblink
core. I want to support faster transfers which is going to require
a few changes. There are also still some send issues every now and
then that pop up.

Signed-off-by: Kris Bahnsen <Kris@KBEmbedded.com>