gblink: strip out ISR in an attempt to support Transfer
Basically, stripping it down to the bare minimum still isn't quite
fast enough to get data consistently. Tucking this away for now.
Likely, in order to support this I'll probably need to implement
using the SPI bus of the flipper.

This could probably also write the ISR in more bare metal, but,
losing access to the FURI API means porting it would be terrible.

Signed-off-by: Kris Bahnsen <Kris@KBEmbedded.com>