gblink: Performance test for measuring interrupt latency
Note that this will also add some amount of latency using the FURI
API for manipulating the GPIO.

Current tests with both standard print as well as with compression
put IRQ latency on either clk edge to:
2.6/19.2/4.4 us (best/worst/average)
This is from clock edge until return from interrupt.

Meaning the latest code can handle a clock up to ~26 kHz with no
loss of data. Some low hanging fruit would be to only IRQ on the
rising edge, read the data on the bus, then set the new data out.
That should be plenty of time to not violate any hold time on the
shift register.

The bottom half handler for both had a run time of:
22.7/192.1/23.9 us (best/worst/average)

This means so long as a byte takes no longer than 192.1 us to shift
in, we shouldn't miss any data. Though this is a bit more complex
since its possible to respond to most bytes with a 0 and be fine and
not have missed any bytes. This would be a max clock rate of roughly
40 kHz. So our limiting factor at this point is the interrupt latency.

Signed-off-by: Kris Bahnsen <Kris@KBEmbedded.com>